EOS-12728: Create separate CortxCli RPM
    * move cli conf files to csm/cli/conf directory
    * handle rsyslog.d for cli
    * small fixes
